# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Library API",
    "description": "A simple library API.",
    "contact": {
      "name": "API Support",
      "url": "http://www.example.com/support",
      "email": "support@example.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/shelves": {
      "get": {
        "summary": "List Shelves",
        "description": "List shelves in the library.",
        "operationId": "ListShelves",
        "parameters": [
          {
            "name": "max_page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Max Page Size"
            }
          },
          {
            "name": "page_token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Page Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListShelvesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Shelf",
        "description": "Create a new shelf.",
        "operationId": "CreateShelf",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Shelf"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shelf"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/shelves/{shelf_id}": {
      "get": {
        "summary": "Get Shelf",
        "description": "Get a shelf by ID.",
        "operationId": "GetShelf",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shelf Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shelf"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Shelf",
        "description": "Delete a shelf.",
        "operationId": "DeleteShelf",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shelf Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Shelf",
        "description": "Update a shelf.",
        "operationId": "UpdateShelf",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shelf Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/merge-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ShelfUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shelf"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/shelves/{shelf_id}/books": {
      "get": {
        "summary": "List Books",
        "description": "List books on a shelf.",
        "operationId": "ListBooks",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shelf Id"
            }
          },
          {
            "name": "max_page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Max Page Size"
            }
          },
          {
            "name": "page_token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Page Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListBooksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Book",
        "description": "Create a new book on a shelf.",
        "operationId": "CreateBook",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shelf Id"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/shelves/{shelf_id}/books/{book_id}": {
      "get": {
        "summary": "Get Book",
        "description": "Get a book by ID.",
        "operationId": "GetBook",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shelf Id"
            }
          },
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Book",
        "description": "Delete a book.",
        "operationId": "DeleteBook",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shelf Id"
            }
          },
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Book",
        "description": "Update a book.",
        "operationId": "UpdateBook",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shelf Id"
            }
          },
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/merge-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/BookUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Book": {
        "properties": {
          "path": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^shelves/[a-zA-Z0-9\\-]+/books/[a-zA-Z0-9\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "type": "string",
            "title": "Path",
            "description": "The resource path.",
            "readOnly": true
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "The title of the book."
          },
          "author": {
            "type": "string",
            "title": "Author",
            "description": "The author of the book."
          }
        },
        "type": "object",
        "required": [
          "title",
          "author"
        ],
        "title": "Book",
        "x-aep-resource": {
          "patterns": [
            "shelves/{shelf}/books/{book}"
          ],
          "plural": "books",
          "singular": "book",
          "type": "library.example.com/book"
        }
      },
      "BookUpdate": {
        "properties": {
          "path": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^shelves/[a-zA-Z0-9\\-]+/books/[a-zA-Z0-9\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "type": "string",
            "title": "Path",
            "description": "The resource path.",
            "readOnly": true
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          }
        },
        "type": "object",
        "title": "BookUpdate",
        "x-aep-resource": {
          "patterns": [
            "shelves/{shelf}/books/{book}"
          ],
          "plural": "books",
          "singular": "book",
          "type": "library.example.com/book"
        }
      },
      "ListBooksResponse": {
        "properties": {
          "books": {
            "items": {
              "$ref": "#/components/schemas/Book"
            },
            "type": "array",
            "title": "Books"
          },
          "next_page_token": {
            "type": "string",
            "title": "Next Page Token",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "books"
        ],
        "title": "ListBooksResponse"
      },
      "ListShelvesResponse": {
        "properties": {
          "shelves": {
            "items": {
              "$ref": "#/components/schemas/Shelf"
            },
            "type": "array",
            "title": "Shelves"
          },
          "next_page_token": {
            "type": "string",
            "title": "Next Page Token",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "shelves"
        ],
        "title": "ListShelvesResponse"
      },
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference",
            "title": "Type",
            "description": "A URI reference that identifies the problem type."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "A short, human-readable summary of the problem type."
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "The HTTP status code generated by the origin server for this occurrence of the problem."
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "type": "string",
            "title": "Detail",
            "description": "A human-readable explanation specific to this occurrence of the problem."
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "type": "string",
            "format": "uri-reference",
            "title": "Instance",
            "description": "A URI reference that identifies the specific occurrence of the problem."
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem Details"
      },
      "Shelf": {
        "properties": {
          "path": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^shelves/[a-zA-Z0-9\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "type": "string",
            "title": "Path",
            "description": "The resource path.",
            "readOnly": true
          },
          "theme": {
            "type": "string",
            "title": "Theme",
            "description": "The theme of the shelf."
          }
        },
        "type": "object",
        "required": [
          "theme"
        ],
        "title": "Shelf",
        "x-aep-resource": {
          "patterns": [
            "shelves/{shelf}"
          ],
          "plural": "shelves",
          "singular": "shelf",
          "type": "library.example.com/shelf"
        }
      },
      "ShelfUpdate": {
        "properties": {
          "path": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^shelves/[a-zA-Z0-9\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "type": "string",
            "title": "Path",
            "description": "The resource path.",
            "readOnly": true
          },
          "theme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme"
          }
        },
        "type": "object",
        "title": "ShelfUpdate",
        "x-aep-resource": {
          "patterns": [
            "shelves/{shelf}"
          ],
          "plural": "shelves",
          "singular": "shelf",
          "type": "library.example.com/shelf"
        }
      }
    }
  }
}